feat: add compatibility setting for non-Windows
Windows has certain 'illegal characters', this means that you cannot use
them in file names, Linux doesn't have this limitation; so there is a
risk that you make a note on Linux with a title that is not compatible
Windows, in this case you wouldn't be able to move over your vault from
a Unix system into your Windows machine. This setting applies the same
checks for note names on your non-Windows machine so that you can retain
compatibility between the two.