feat: add ability to force create a note
There are certain cases where you want to create a note with a specific
title, but the title is similar to a template, so when you try to create
it you will use the template even if you don't want to. With this now
you can use `ctrl + enter` to 'force' the creation of the note, it works
just like when you don't have any template selected/when the templates
are deactivated.