feat: add ability to override default 'new note'
With this new commit you now can override the functionality of the
'New note' button that is located on top of the file explorer, if you
only activate the option to override the functionality without selecting
a command, the program will default to call the command associated with
'New tab'. Select any other mode of creation via the dropdown menu that
is underneath the slider to override the functionality.