feat: add lables/renderer for note and title templates
Now when you go to create a note with templates, the plugin will show
you if you are choosing a Note Template or a Title Template, you can
choose both for the same note. So both the title of your note and the
body of your note will be based on a template, but you need to choose
first the note template which will then let you choose a title template,
this is because if you choose first a title template the plugin assumes
that you don't want a note template and just creates the note with the
title you gave it via the title template.